clean up allocator interface to be simpler
