+++ goalc/compiler/Compiler.h
@@ -79,6 +79,10 @@ class Compiler {
   std::unique_ptr<ReplWrapper> m_repl;
   MakeSystem m_make;
 
+  struct DebugStats {
+    int num_spills = 0;
+  } m_debug_stats;
+
   std::set<std::string> lookup_symbol_infos_starting_with(const std::string& prefix) const;
   std::vector<SymbolInfo>* lookup_exact_name_info(const std::string& name) const;
   bool get_true_or_false(const goos::Object& form, const goos::Object& boolean);
