+++ goalc/compiler/IR.h
@@ -2,7 +2,7 @@
 
 #include <string>
 #include "CodeGenerator.h"
-#include "goalc/regalloc/allocate.h"
+#include "goalc/regalloc/allocator_interface.h"
 #include "Val.h"
 #include "goalc/emitter/ObjectGenerator.h"
 #include "goalc/emitter/Register.h"
