+++ goalc/regalloc/Allocator_v2.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "goalc/regalloc/allocator_interface.h"
+
+// Allocator v2's interface
+AllocationResult allocate_registers_v2(const AllocationInput& input);
